fixes to preprocessing according to thesis and adjustments to measurement system

+++ task.h
@@ -202,8 +202,8 @@ typedef struct __attribute__((__packed__)) RealTimeTask_t {
 	u32 pxCriticalityLevel;
 } RTTask_t;
 
-void generate_deadlines(u32 tasksDerivativesDeadlines[configCRITICALITY_LEVELS][configMAX_RT_TASKS], u32 tasksDeadlines[configCRITICALITY_LEVELS][configMAX_RT_TASKS], RTTask_t task, int taskIndex, u32 x, u32 k);
-int calculate_x(RTTask_t tasks[], u8 numberOfTasks, int k);
+void generate_deadlines(u32 tasksDerivativesDeadlines[configCRITICALITY_LEVELS][configMAX_RT_TASKS], u32 tasksDeadlines[configCRITICALITY_LEVELS][configMAX_RT_TASKS], RTTask_t task, int taskIndex, float x, u32 k);
+float calculate_x(RTTask_t tasks[], u8 numberOfTasks, int k);
 int find_k(RTTask_t tasks[], u8 number_of_tasks);
 float compute_utilisation(RTTask_t tasks[], u8 number_of_tasks, u32 systemCriticalityLevel, u32 taskCriticalityLevel);
 int prvGenerateSchedulerDataFromTaskSet(RTTask_t prvRTTasksList[configMAX_RT_TASKS], u8 numberOfTasks,
